Move versions to internal/versions
versions (the versions.env reader) has no consumer outside the root module.
Application code that isn't part of a public API belongs under internal/,
where the import boundary is compiler-enforced.

versions.go locates versions.env at runtime by walking up from its own source
path (runtime.Caller), so the added directory level requires one more
filepath.Dir to still reach the repo root. Verified at runtime that
versions.env still resolves after the move.

stack-info: PR: https://github.com/temporalio/omes/pull/375, branch: lilydoar/stack/9
