+++ dockerfiles/typescript.Dockerfile
@@ -26,7 +26,6 @@ COPY clioptions ./clioptions
 COPY loadgen ./loadgen
 COPY scenarios ./scenarios
 COPY metrics ./metrics
-COPY versions ./versions
 COPY internal ./internal
 COPY workers/go/harness/api ./workers/go/harness/api
 COPY go.mod go.sum versions.env ./
